Separate the first statistic stuff
